Enable tracking of k-best or within-threshold-from-best parameters
Storing parameters for all evaluated points during optimization can become costly for larger problems.
Often, we are only interested in the top-k parameter vectors, or the ones within a certain fval-distance from the best one.
This adds the respective filters.

Auto-save functionality is still missing. Nothing will be stored if the run is interrupted.

Related to #13.